Cast the requested time to a double to make the division work. The dynamic timer works now for periods > 100 seconds.
svn path=/trunk/playground/utils/rsibreak/; revision=487778

+++ src/rsitimer.cpp
@@ -246,7 +246,7 @@ void RSITimer::timerEvent( QTimerEvent* )
     }
     else //active
     {
-      idleIndex += (100 / m_intervals["time_minimized"]);
+      idleIndex += 100 / (double)(m_intervals["time_minimized"]);
     }
 
     emit updateIdleAvg( idleIndex );
